% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_data.R
\docType{data}
\name{coffee_data}
\alias{coffee_data}
\title{Sample dataset from a course about analysis of factorial designs}
\description{
A sample data set from a course about the analysis of factorial
designs, by Mattan S. Ben-Shachar. See following link for more information:
https://github.com/mattansb/Analysis-of-Factorial-Designs-foR-Psychologists

The data consists of five variables from 120 observations:
\itemize{
\item \code{ID}: A unique identifier for each participant
\item \code{sex}: The participant's sex
\item \code{time}: The time of day the participant was tested (morning or afternoon)
\item \code{coffee}: Group indicator, whether participant drank coffee or not
("\verb{coffee"} or \code{"control"}).
\item \code{alertness}: The participant's alertness score.
}
}
\examples{
# Attach coffee-data
data(coffee_data)
}
\keyword{data}
