% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{comparisons}
\alias{comparisons}
\title{(Pairwise) comparisons between predictions}
\usage{
comparisons(x, test = "pairwise")
}
\arguments{
\item{x}{A \code{ggeffects} object.}

\item{test}{Hypothesis to test. By default, pairwise-comparisons are conducted.}
}
\value{
A data frame containing...
}
\description{
Create...
}
\examples{
data(efc)
efc$c172code <- as.factor(efc$c172code)
m <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
pred <- ggpredict(m, "c172hour")
comparisons(pred)
}
