% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.R
\docType{data}
\name{efc}
\alias{efc}
\alias{efc_test}
\title{Sample dataset from the EUROFAMCARE project}
\description{
A SPSS sample data set, imported with the \code{sjlabelled::read_spss()} function.
}
\examples{
# Attach EFC-data
data(efc)

# Show structure
str(efc)

# show first rows
head(efc)
}
\keyword{data}
