% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Sample data set}
\description{
A sample data set, used in tests and some examples. Useful for
demonstrating count models (with or without zero-inflation component). It
consists of nine variables from 250 observations.
}
\keyword{data}
