% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparisons.R
\name{ggcomparisons}
\alias{ggcomparisons}
\alias{ggcomparisons.default}
\alias{ggcomparisons.ggeffects}
\title{(Pairwise) comparisons between predictions}
\usage{
ggcomparisons(model, ...)

\method{ggcomparisons}{default}(model, terms = NULL, test = "pairwise", ...)

\method{ggcomparisons}{ggeffects}(model, test = "pairwise", ...)
}
\arguments{
\item{model}{A fitted model object, or an object of class \code{ggeffects}.}

\item{...}{Currently not used.}

\item{terms}{Character vector with the names of the focal terms from \code{model},
for which contrasts or comparisons should be displayed. At least one term
is required, maximum length is three terms. One focal term may be numeric.
In this case, contrasts or comparisons for the \emph{slopes} of the numeric
predictor are computed (possibly grouped by the levels of further categorical
focal predictors).}

\item{test}{Hypothesis to test. By default, pairwise-comparisons are conducted.}
}
\value{
A data frame containing...
}
\description{
Create...
}
\examples{
if (requireNamespace("marginaleffects")) {
  data(efc)
  efc$c172code <- as.factor(efc$c172code)
  efc$c161sex <- as.factor(efc$c161sex)
  levels(efc$c161sex) <- c("male", "female")
  m <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)

  # direct computation of comparisons
  ggcomparisons(m, "c172code")

  # passing a `ggeffects` object
  pred <- ggpredict(m, "c172code")
  ggcomparisons(pred)

  # test for slope
  ggcomparisons(m, "c12hour")

  # interaction - contrasts by groups
  m <- lm(barthtot ~ c12hour + c161sex * c172code + neg_c_7, data = efc)
  ggcomparisons(m, c("c161sex", "c172code"), test = NULL)

  # interaction - pairwise comparisons by groups
  ggcomparisons(m, c("c161sex", "c172code"))

  # specific comparisons
  ggcomparisons(m, c("c161sex", "c172code"), test = "b2 = b1")

  # interaction - slope by groups
  m <- lm(barthtot ~ c12hour + neg_c_7 * c172code + c161sex, data = efc)
  ggcomparisons(m, c("neg_c_7", "c172code"))
}
}
