% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnson_neyman.R
\name{johnson_neyman}
\alias{johnson_neyman}
\title{Create Johnson-Neyman plot}
\usage{
johnson_neyman(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggeffects}, as returned by the functions
from this package.}

\item{...}{Currently not used.}
}
\description{
Function to create so-called Johnson-Neyman plots, which are
used to visualize the results of a Johnson-Neyman test.
}
