% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{lung}
\alias{lung}
\title{Sample data set}
\description{
A sample data set, used in tests and examples for survival models.
This dataset is originally included in the \pkg{survival} package, but for
convenience reasons it is also available in this package.
}
\keyword{data}
