% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov}
\alias{vcov}
\alias{vcov.ggeffects}
\title{Calculate variance-covariance matrix for marginal effects}
\usage{
\method{vcov}{ggeffects}(object, vcov.fun = NULL, vcov.type = NULL,
  vcov.args = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{"ggeffects"}, as returned by \code{ggpredict()}.}

\item{vcov.fun}{String, indicating the name of the \code{vcov*()}-function
from the \pkg{sandwich}-package, e.g. \code{vcov.fun = "vcovCL"},
which is used to compute robust standard errors for predictions.
If \code{NULL}, standard errors (and confidence intervals) for predictions
are based on the standard errors as returned by the \code{predict()}-function.
\strong{Note} that probably not all model objects that work with \code{ggpredict()}
are also supported by the \pkg{sandwich}-package.}

\item{vcov.type}{Character vector, specifying the estimation type for the
robust covariance matrix estimation (see \code{\link[sandwich]{vcovHC}}
for details).}

\item{vcov.args}{List of named vectors, used as additional arguments that
are passed down to \code{vcov.fun}.}

\item{...}{Currently not used.}
}
\value{
The variance-covariance matrix for the predicted values from \code{object}.
}
\description{
Returns the variance-covariance matrix for the predicted values from \code{object}.
}
\examples{
data(efc)
model <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
result <- ggpredict(model, c("c12hour [meansd]", "c161sex"))

vcov(result)

# compare standard errors
sqrt(diag(vcov(result)))
as.data.frame(result)

}
