% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_data.R
\docType{data}
\name{coffee_data}
\alias{coffee_data}
\title{Sample dataset}
\description{
A sample data set from a course about analysis of factorial
Designs, by Mattan S. Ben-Shachar. See following link for more information:
https://github.com/mattansb/Analysis-of-Factorial-Designs-foR-Psychologists
}
\examples{
# Attach coffee-data
data(coffee_data)
}
\keyword{data}
