% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efc.R
\docType{data}
\name{efc}
\alias{efc}
\alias{efc_test}
\title{Sample dataset from the EUROFAMCARE project}
\description{
An SPSS sample data set, imported with the \code{\link[sjlabelled:read_spss]{sjlabelled::read_spss()}}
function. Consists of 28 variables from 908 observations. The data set is part
of the EUROFAMCARE project, a cross-national survey on informal caregiving in
Europe.
}
\examples{
# Attach EFC-data
data(efc)

# Show structure
str(efc)

# show first rows
head(efc)
}
\keyword{data}
