% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnson_neyman.R
\name{johnson_neyman}
\alias{johnson_neyman}
\title{Create Johnson-Neyman plot}
\usage{
johnson_neyman(x, ...)
}
\arguments{
\item{x}{An object of class \code{ggeffects}, as returned by the functions
from this package.}

\item{...}{Arguments passed down to \code{hypothesis_test()}.}
}
\value{
A Johnson-Neyman plot.
}
\description{
Function to create so-called Johnson-Neyman plots, which are
used to visualize the results of a Johnson-Neyman test.
}
\examples{
data(efc)
efc$c172code <- as.factor(efc$c172code)
m <- lm(neg_c_7 ~ c12hour * barthtot * c172code, data = efc)

if (requireNamespace("ggplot2") &&
    requireNamespace("see") &&
    requireNamespace("marginaleffects")) {
  pr <- ggpredict(m, c("c12hour", "barthtot"))
  johnson_neyman(pr)

  pr <- ggpredict(m, c("c12hour", "c172code", "barthtot"))
  johnson_neyman(pr)

  # robust standard errors
  if (requireNamespace("sandwich")) {
    johnson_neyman(pr, vcov = sandwich::vcovHC)
  }
}
}
