% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R, R/print.R
\name{format.ggeffects}
\alias{format.ggeffects}
\alias{print}
\alias{print.ggeffects}
\title{Print and format ggeffects-objects}
\usage{
\method{format}{ggeffects}(
  x,
  variable_labels = FALSE,
  value_labels = FALSE,
  group_name = FALSE,
  row_header_separator = ", ",
  n,
  ...
)

\method{print}{ggeffects}(x, group_name = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{ggeffects}, as returned by the functions
from this package.}

\item{...}{Further arguments passed down to \verb{ggplot::scale_y*()}, to
control the appearance of the y-axis.}
}
\value{
A ggplot2-object.
}
\description{
A generic plot-method for \code{ggeffects}-objects.
}
\examples{
data(efc)
efc$c172code <- as_label(efc$c172code)
fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)

dat <- ggpredict(fit, terms = "c12hour")
plot(dat)

\donttest{
# facet by group, use pre-defined color palette
dat <- ggpredict(fit, terms = c("c12hour", "c172code"))
plot(dat, facet = TRUE, colors = "hero")

# don't use facets, b/w figure, w/o confidence bands
dat <- ggpredict(fit, terms = c("c12hour", "c172code"))
plot(dat, colors = "bw", show_ci = FALSE)

# factor at x axis, plot exact data points and error bars
dat <- ggpredict(fit, terms = c("c172code", "c161sex"))
plot(dat)

# for three variables, automatic facetting
dat <- ggpredict(fit, terms = c("c12hour", "c172code", "c161sex"))
plot(dat)
}

# show all color palettes
show_pals()
}
